Adding InteractionOptions sample. Modified existing 3D samples category to Scene.
