+++ src/Desktop/ArcGISRuntimeSamplesDesktop/ArcGISRuntime.Samples.Desktop.csproj
@@ -31,6 +31,7 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <DocumentationFile>Assets\SampleDescriptions.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="Esri.ArcGISRuntime, Version=10.2.6.0, Culture=neutral, PublicKeyToken=8fc3cc631e44ad86, processorArchitecture=MSIL" />
@@ -55,24 +56,27 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Samples\3D\ArcGISDynamicMapServiceLayerSample3d.xaml.cs">
+    <Compile Include="Samples\Scene\ArcGISDynamicMapServiceLayerSample3d.xaml.cs">
       <DependentUpon>ArcGISDynamicMapServiceLayerSample3d.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Samples\3D\ArcGISTiledMapServiceLayerSample3d.xaml.cs">
+    <Compile Include="Samples\Scene\ArcGISTiledMapServiceLayerSample3d.xaml.cs">
       <DependentUpon>ArcGISTiledMapServiceLayerSample3d.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Samples\3D\CameraAnimationSample3d.xaml.cs">
+    <Compile Include="Samples\Scene\CameraAnimationSample3d.xaml.cs">
       <DependentUpon>CameraAnimationSample3d.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Samples\3D\ElevationSourceSample3d.xaml.cs">
+    <Compile Include="Samples\Scene\ElevationSourceSample3d.xaml.cs">
       <DependentUpon>ElevationSourceSample3d.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Samples\3D\ElevationModeSample.xaml.cs">
+    <Compile Include="Samples\Scene\ElevationModeSample.xaml.cs">
       <DependentUpon>ElevationModeSample.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Samples\3D\OpenStreetMapLayerSample3d.xaml.cs">
+    <Compile Include="Samples\Scene\InteractionOptionsSample3d.xaml.cs">
+      <DependentUpon>InteractionOptionsSample3d.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Samples\Scene\OpenStreetMapLayerSample3d.xaml.cs">
       <DependentUpon>OpenStreetMapLayerSample3d.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\DynamicServiceLayers\ArcGISDynamicMapServiceLayerLocalSample.xaml.cs">
@@ -502,27 +506,31 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Page Include="Samples\3D\ArcGISDynamicMapServiceLayerSample3d.xaml">
+    <Page Include="Samples\Scene\ArcGISDynamicMapServiceLayerSample3d.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Samples\3D\ArcGISTiledMapServiceLayerSample3d.xaml">
+    <Page Include="Samples\Scene\ArcGISTiledMapServiceLayerSample3d.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Samples\3D\CameraAnimationSample3d.xaml">
+    <Page Include="Samples\Scene\CameraAnimationSample3d.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Samples\3D\ElevationSourceSample3d.xaml">
+    <Page Include="Samples\Scene\ElevationSourceSample3d.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Samples\3D\ElevationModeSample.xaml">
+    <Page Include="Samples\Scene\ElevationModeSample.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Samples\3D\OpenStreetMapLayerSample3d.xaml">
+    <Page Include="Samples\Scene\InteractionOptionsSample3d.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Samples\Scene\OpenStreetMapLayerSample3d.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -1092,7 +1100,9 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\SampleDescriptions.xml" />
+    <EmbeddedResource Include="Assets\SampleDescriptions.xml">
+      <SubType>Designer</SubType>
+    </EmbeddedResource>
   </ItemGroup>
   <ItemGroup />
   <ItemGroup>
