--- src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/3D/ArcGISTiledMapServiceLayerSample3d.xaml
@@ -1,18 +1,0 @@
-﻿<UserControl 
-	x:Class="ArcGISRuntime.Samples.Desktop.ArcGISTiledMapServiceLayerSample3d"
-	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-	xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-	xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
-	mc:Ignorable="d" 
-	d:DesignHeight="300" d:DesignWidth="300">
-	<Grid>
-		<esri:SceneView x:Name="MySceneView">
-			<esri:Scene>
-				<esri:ArcGISTiledMapServiceLayer ID="StreetMap"
-					ServiceUri="http://services.arcgisonline.com/arcgis/rest/services/World_Street_Map/MapServer" />
-			</esri:Scene>
-		</esri:SceneView>
-	</Grid>
-</UserControl>
