--- src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/3D/ArcGISTiledMapServiceLayerSample3d.xaml.cs
@@ -1,16 +1,0 @@
-﻿namespace ArcGISRuntime.Samples.Desktop
-{
-	/// <summary>
-	/// Demonstrates adding an 2d ArcGIS tiled map service to a Scene in XAML.
-	/// </summary>
-	/// <title>3D ArcGIS Tiled Map Service Layer</title>
-	/// <category>3D</category>
-	/// <subcategory>Layers</subcategory>
-	public partial class ArcGISTiledMapServiceLayerSample3d
-	{
-		public ArcGISTiledMapServiceLayerSample3d()
-		{
-			InitializeComponent();
-		}
-	}
-}
