--- src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/3D/GraphicsExtrusionSample3d.xaml
@@ -1,77 +1,0 @@
-﻿<UserControl 
-	x:Class="ArcGISRuntime.Samples.Desktop.GraphicsExtrusionSample3d"
-	  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-			 xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
-			 xmlns:local="clr-namespace:ArcGISRuntime.Samples.Desktop"
-			 mc:Ignorable="d" 
-			 d:DesignHeight="300" d:DesignWidth="300">
-	<UserControl.Resources>
-		<local:ExtrusionGraphicsMainViewModel x:Key="VM">
-			<local:ExtrusionGraphicsMainViewModel.Scene>
-				<esri:Scene>
-					<esri:Scene.Surface>
-						<esri:ElevationSource ServiceUri="http://elevation3d.arcgis.com/arcgis/rest/services/WorldElevation3D/Terrain3D/ImageServer"
-                                      IsEnabled="True" />
-					</esri:Scene.Surface>
-					<esri:ArcGISTiledMapServiceLayer ID="AGOLayer"
-                ServiceUri="http://services.arcgisonline.com/ArcGIS/rest/services/World_Street_Map/MapServer" />
-					<esri:GraphicsLayer  ID="DynamicGraphicsLayer" RenderingMode="Dynamic" IsVisible="True">
-						<esri:GraphicsLayer.SceneProperties>
-							<esri:LayerSceneProperties SurfacePlacement="Draped"/>
-						</esri:GraphicsLayer.SceneProperties>
-					</esri:GraphicsLayer>
-				</esri:Scene>
-			</local:ExtrusionGraphicsMainViewModel.Scene>
-		</local:ExtrusionGraphicsMainViewModel>
-	</UserControl.Resources>
-	<Grid DataContext="{StaticResource VM}">
-		<Grid.Resources>
-
-		</Grid.Resources>
-		<esri:SceneView x:Name="MySceneView" Scene="{Binding Path=Scene}"/>
-		<StackPanel HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10" Orientation="Horizontal" Background="White">
-			<RadioButton IsChecked="False" Margin="5" 
-                         GroupName="DrawShapes" Content="Add Points" Command="{Binding AddShapesCommand}" CommandParameter="Point"/>
-			<RadioButton IsChecked="False" Margin="5" 
-                         GroupName="DrawShapes" Content="Add Polyline" Command="{Binding AddShapesCommand}" CommandParameter="Polyline"/>
-			<RadioButton IsChecked="False" Margin="5" 
-                         GroupName="DrawShapes" Content="Add Polygon" Command="{Binding AddShapesCommand}" CommandParameter="Polygon"/>
-			<!--<Button Content="Add Points" Margin="5" Command="{Binding AddShapesCommand}" CommandParameter="Point"/>
-            <Button Content="Add Polylines" Margin="5" Command="{Binding AddShapesCommand}" CommandParameter="Polyline"/>
-            <Button Content="Add Polygons" Margin="5" Command="{Binding AddShapesCommand}" CommandParameter="Polygon"/>
-            <Button Content="Clear All" Margin="5" Command="{Binding ClearAllCommand}"/>-->
-
-			<!--<Button Content="Camera" Click="Button_Click"/>-->
-			<!--<RadioButton x:Name="AttributeAExtrusionRadioButton" 
-                         Command="{Binding SwitchExtrusionAttributeCommand}" CommandParameter="A"
-                         Tag="A"
-                         IsChecked="true" Margin="5" 
-                         GroupName="Extrusion" Content="Extrude on A" />
-            <RadioButton x:Name="AttributeBExtrusionRadioButton" 
-                         Command="{Binding SwitchExtrusionAttributeCommand}" CommandParameter="B"
-                         Tag="B" 
-                         Margin="5" 
-                         GroupName="Extrusion" Content="Extrude on B" />-->
-		</StackPanel>
-
-		<StackPanel HorizontalAlignment="Right" VerticalAlignment="Top" Margin="10" Orientation="Horizontal" Background="White">
-			<TextBlock TextAlignment="Center" Text="Select the extrusion mode" 
-                      FontWeight="Bold" Height="16" Margin="5" />
-			<ComboBox Margin="5" ItemsSource="{Binding ExtrusionRenderers}" SelectedItem="{Binding SelectedExtrusionRenderer, Mode=TwoWay}" IsEnabled="{Binding IsGraphicAvailable}" DisplayMemberPath="Title"/>
-
-			<RadioButton x:Name="AttributeAExtrusionRadioButton" 
-                         Command="{Binding SwitchExtrusionAttributeCommand}" CommandParameter="A"
-                         IsChecked="true" Margin="5" 
-                         IsEnabled="{Binding IsGraphicAvailable}"
-                         GroupName="Extrusion" Content="Extrude on Value 1" />
-			<RadioButton x:Name="AttributeBExtrusionRadioButton" 
-                         Command="{Binding SwitchExtrusionAttributeCommand}" CommandParameter="B"
-                         IsEnabled="{Binding IsGraphicAvailable}"
-                         Margin="5" 
-                         GroupName="Extrusion" Content="Extrude on Value 2" />
-		</StackPanel>
-	</Grid>
-</UserControl>
