--- src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/3D/OpenStreetMapLayerSample3d.xaml.cs
@@ -1,18 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace ArcGISRuntime.Samples.Desktop
-{
-	/// <summary>
-	/// Demonstrates adding an OpenStreetMap to a Scene in XAML.
-	/// </summary>
-	/// <title>3D Open Street map</title>
-	/// <category>3D</category>
-	/// <subcategory>Layers</subcategory>
-	public partial class OpenStreetMapLayerSample3d : UserControl
-	{
-		public OpenStreetMapLayerSample3d()
-		{
-			InitializeComponent();
-		}
-	}
-}
