+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/ArcGISDynamicMapServiceLayerSample3d.xaml
@@ -1,0 +1,20 @@
+﻿<UserControl 
+	x:Class="ArcGISRuntime.Samples.Desktop.ArcGISDynamicMapServiceLayerSample3d"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+	xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
+	mc:Ignorable="d" 
+	d:DesignHeight="300" d:DesignWidth="300">
+	<Grid>
+		<esri:SceneView x:Name="MySceneView">
+			<esri:Scene>
+				<esri:ArcGISTiledMapServiceLayer ID="Imagery"
+					ServiceUri="http://services.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer" />
+				<esri:ArcGISDynamicMapServiceLayer 
+					ServiceUri="http://sampleserver6.arcgisonline.com/arcgis/rest/services/USA/MapServer"/>
+			</esri:Scene>
+		</esri:SceneView>
+	</Grid>
+</UserControl>
