+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/ArcGISTiledMapServiceLayerSample3d.xaml.cs
@@ -1,0 +1,16 @@
+﻿namespace ArcGISRuntime.Samples.Desktop
+{
+	/// <summary>
+	/// Demonstrates adding an 2d ArcGIS tiled map service to a Scene in XAML.
+	/// </summary>
+	/// <title>3D ArcGIS Tiled Map Service Layer</title>
+	/// <category>Scene</category>
+	/// <subcategory>Layers</subcategory>
+	public partial class ArcGISTiledMapServiceLayerSample3d
+	{
+		public ArcGISTiledMapServiceLayerSample3d()
+		{
+			InitializeComponent();
+		}
+	}
+}
