+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/ElevationSourceSample3d.xaml
@@ -1,0 +1,22 @@
+﻿<UserControl 
+	x:Class="ArcGISRuntime.Samples.Desktop.ElevationSourceSample3d"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+	xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
+	mc:Ignorable="d" 
+	d:DesignHeight="300" d:DesignWidth="300">
+	<Grid>
+		<esri:SceneView x:Name="MySceneView">
+			<esri:Scene>
+				<esri:Scene.Surface>
+					<esri:ServiceElevationSource IsEnabled="True" 
+						ServiceUri="http://elevation3d.arcgis.com/arcgis/rest/services/WorldElevation3D/Terrain3D/ImageServer"/>
+				</esri:Scene.Surface>
+				<esri:ArcGISTiledMapServiceLayer ID="Imagery"
+					ServiceUri="http://services.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer" />
+			</esri:Scene>
+		</esri:SceneView>
+	</Grid>
+</UserControl>
