+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/InteractionOptionsSample3d.xaml
@@ -1,0 +1,95 @@
+﻿<UserControl 
+	x:Class="ArcGISRuntime.Samples.Desktop.InteractionOptionsSample3d"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:sys="clr-namespace:System;assembly=mscorlib"
+    xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
+    <Grid>
+		<esri:SceneView 
+				x:Name="MySceneView"
+				LayerLoaded="MySceneView_LayerLoaded">
+			<!-- Setting values explicitly -->
+			<!--<esri:SceneView.InteractionOptions>
+				<esri:InteractionOptions IsEnabled="True">
+					<esri:InteractionOptions.PanOptions>
+						<esri:PanOptions
+							IsEnabled="True"
+							IsDragEnabled="True"
+							IsFlickEnabled="True"
+							IsKeyboardEnabled="True"
+							IsStylusEnabled="True">
+						</esri:PanOptions>
+					</esri:InteractionOptions.PanOptions>
+				</esri:InteractionOptions>			
+			</esri:SceneView.InteractionOptions>-->
+			<esri:Scene>
+				<esri:ArcGISTiledMapServiceLayer ID="Basemap"
+							ServiceUri="http://services.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer"/>
+			</esri:Scene>
+		</esri:SceneView>
+
+			<Border Background="White" BorderBrush="Black" BorderThickness="1" Margin="30" Padding="20"
+							HorizontalAlignment="Right" VerticalAlignment="Top">
+			<Border.Effect>
+				<DropShadowEffect />
+			</Border.Effect>
+				<StackPanel>
+					<TextBlock Text="Interaction options" HorizontalAlignment="Center" FontSize="16" Margin="10" FontWeight="Bold"/>
+					<StackPanel Margin="10,0">
+					<CheckBox Content="IsEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.IsEnabled}"/>
+					</StackPanel>
+
+				<StackPanel Margin="10,0">
+					<TextBlock Text="Pan options" HorizontalAlignment="Left" FontSize="12" Margin="0,10" FontWeight="Bold"/>
+					<CheckBox Content="IsEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.PanOptions.IsEnabled}"/>
+					<CheckBox Content="IsDragEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.PanOptions.IsDragEnabled}"/>
+					<CheckBox Content="IsFlickEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.PanOptions.IsFlickEnabled}"/>
+					<CheckBox Content="IsKeyboardEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.PanOptions.IsKeyboardEnabled}"/>
+					<CheckBox Content="IsStylusEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.PanOptions.IsStylusEnabled}"/>
+				</StackPanel>
+
+				<StackPanel Margin="10,0">
+					<StackPanel.Resources>
+						<ObjectDataProvider x:Key="MouseWheelOptionValues" MethodName="GetValues" ObjectType="{x:Type sys:Enum}">
+							<ObjectDataProvider.MethodParameters>
+								<x:Type TypeName="esri:MouseWheelDirection"/>
+							</ObjectDataProvider.MethodParameters>
+						</ObjectDataProvider>
+					</StackPanel.Resources>
+					<TextBlock Text="Zoom options" HorizontalAlignment="Left" FontSize="12" Margin="0,10" FontWeight="Bold"/>
+					<CheckBox Content="IsEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsEnabled}"/>
+					<CheckBox Content="IsDoubleTappedEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsDoubleTappedEnabled}"/>
+					<CheckBox Content="IsKeyboardEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsKeyboardEnabled}"/>
+					<CheckBox Content="IsMouseWheelEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsMouseWheelEnabled}"/>
+					<CheckBox Content="IsPinchEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsPinchEnabled}"/>
+					<CheckBox Content="IsTwoFingerTapEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsTwoFingerTapEnabled}"/>
+					<CheckBox Content="IsZoomBoxEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsZoomBoxEnabled}"/>
+					<CheckBox Content="IsStylusEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.IsStylusEnabled}"/>
+
+					<StackPanel Margin="0,5">
+						<TextBlock Text="MouseWheelZoomInDirection"/>
+						<ComboBox 
+						Margin="10,2"
+						ItemsSource="{Binding Source={StaticResource MouseWheelOptionValues}}"
+						SelectedItem="{Binding ElementName=MySceneView, Path=InteractionOptions.ZoomOptions.MouseWheelZoomInDirection}">
+						</ComboBox>
+					</StackPanel>
+				</StackPanel>
+
+				<StackPanel Margin="10,0">
+					<TextBlock Text="Rotation options" HorizontalAlignment="Left" FontSize="12" Margin="0,10" FontWeight="Bold"/>
+					<CheckBox Name="RotationEnabled" Content="IsEnabled" IsChecked="{Binding ElementName=MySceneView, Path=InteractionOptions.RotationOptions.IsEnabled}" Margin="0,0,0,5"/>
+
+					<TextBlock Text="{Binding ElementName=rotationSlider, Path=Value, StringFormat='Rotation: {0:0.0}'}" 
+                         Margin="4,0"  FontWeight="Bold" />
+
+					<Slider x:Name="rotationSlider" Margin="4" ValueChanged="rotationSlider_ValueChanged"
+                  Orientation="Horizontal" Width="100" IsSnapToTickEnabled="True" Visibility="{Binding Path=IsChecked, ElementName=RotationEnabled, Converter={StaticResource BooleanToVisibilityConverter}}" 
+									Value="{Binding Rotation, ElementName=MySceneView, Mode=OneWay, Delay=50}"
+                  TickPlacement="BottomRight" TickFrequency="5"
+                  Minimum="0" Maximum="360" />
+				</StackPanel>
+			</StackPanel>
+		</Border>
+	</Grid>
+</UserControl>
