+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/OpenStreetMapLayerSample3d.xaml
@@ -1,0 +1,17 @@
+﻿<UserControl 
+	x:Class="ArcGISRuntime.Samples.Desktop.OpenStreetMapLayerSample3d"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+	xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
+	mc:Ignorable="d" 
+	d:DesignHeight="300" d:DesignWidth="300">
+	<Grid>
+		<esri:SceneView x:Name="MySceneView">
+			<esri:Scene>
+				<esri:OpenStreetMapLayer ID="OpenStreetMap" />
+			</esri:Scene>
+		</esri:SceneView>
+	</Grid>
+</UserControl>
