+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/OpenStreetMapLayerSample3d.xaml.cs
@@ -1,0 +1,18 @@
+﻿using System.Windows.Controls;
+
+namespace ArcGISRuntime.Samples.Desktop
+{
+	/// <summary>
+	/// Demonstrates adding an OpenStreetMap to a Scene in XAML.
+	/// </summary>
+	/// <title>3D Open Street map</title>
+	/// <category>Scene</category>
+	/// <subcategory>Layers</subcategory>
+	public partial class OpenStreetMapLayerSample3d : UserControl
+	{
+		public OpenStreetMapLayerSample3d()
+		{
+			InitializeComponent();
+		}
+	}
+}
